pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nex-ai"
version = "0.1.0"
description = "The coding agent that remembers — AI coding assistant with persistent memory and error learning."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Nex AI Contributors" }]
keywords = ["ai", "coding-agent", "cli", "anthropic", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "anthropic>=0.40.0",
    "typer[all]>=0.15.0",
    "rich>=13.9.0",
    "httpx>=0.28.0",
    "gitpython>=3.1.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
nex = "nex.cli:app"

[project.urls]
Homepage = "https://github.com/nex-ai/nex-ai"
Repository = "https://github.com/nex-ai/nex-ai"

[tool.hatch.build.targets.wheel]
packages = ["src/nex"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", "F", "W", "I", "UP", "N", "S", "B", "A", "C4",
    "SIM", "TCH", "ICN", "PIE", "PT", "RSE", "RET", "SLF",
    "ARG", "ERA", "PGH", "PLC", "PLE", "PLR", "PLW", "TRY",
    "FLY", "PERF", "RUF",
]
ignore = ["S101", "TRY003", "S603", "S607"]

[tool.mypy]
strict = true
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
